Refactoring Utility Stuff
Nearly Everything is refactored except the games.

Themes is almost done. I want to find a way to display a menu of available themes for a user to chose from when creating a user. Make it more dynamic. This way I can add more themes to the themes list and it'll all load automatically.

After that I'm going to do a few cycles through what I've refactored to look for more opportunities to refactor, rename, and organize the code.