Created DeleteThisUser
It's a menu option in User Menu. It deletes the current user, then calls the log in function.

Problem is, after logging in, it stays in the user menu.

I think, for the sake of extendability, that I need to make a Program class function to completely rerun the program upon deleting a program.