Broke away from Hangman
I had a realization about CodeCracker

There is a very finite number of possible codes.

So instead of randomly generating codes and hints and risking errors, I writting a small program to generate all possible 3 digit codes with no repeating numbers and 5 hints for each. This data will be saved to a text file which will be imported here.

Everything about said new program is finished except for generating the hints. SAving that for tomorrow. One all of the possible codes are generated, I'll be refactoring Code Cracker. Saving Hangman for last. Then War. Then more refacotirng and code cleanup