Code Cracker data file
I wrote and executed a program that created a data file for my Code Cracker to read from. This frees up resources by allowing the game to run without having to randomly generate codes and hints. Instead it just needs to randomly choose one from the pre-created list.

The program I created added every number between 12 and 987 with no repeating digits to a list. It then created 5 hint codes for each code to solve . These were formatted and added to the file. It needs to be improved to avoid similarities between hints for a code to solve. But I currently have a usable file of 720 codes and 3600 hints ( 5 each) for the game to use.