Cleansed bool loops from App Class
I also refactored the Init method to create all the needed menus at once. Each menu method just displays it. Much cleaner and easier to understnad.

Had to create public properties in the menu class to make this work. Just getters for the title of the menu and the number of items in that menu.

Need to take my upgraded menu class from my last assignment and integrate into this program. Then another round of refactorings throughout the program.