Fixed Many a Bug
But now I have issues with backing out of menus.

I also notice a logically flawed bag of shit that is theme selection. Need to rework it to include error messages if the current theme is chosen. It also needs to not display anything if the back button is chosen.