Split Game Menus
The singular game menu is now split into two game menus: 1 player and 2 player.

Going to add functionality to log in a second player and track progress for two player games. They will be loggen out when the game ends and the original user returns to the 2 player games menu.