User Point Tracking
I creating an overall point tracking system that awards the logged in player a point every time they win a game. There is an AddPoint method in user class that can hold some user output in the future as well. Going to use this to add two player functionality on a game by game basis.