Point System Upgraded
I am now tracking not only the overall user score of total wins, but the total wins for each game (a breakdown).

This refactoring has been done in User class, FileIO class, and withiin each game. The User AddPoint method now updates the total and the individual game score based on a string passed in by the game. The string is the key selector for changing values in the dictionary used for scores.

Need to flesh out my display methods and then test extensively.