Added Card Game, War
This is a two player game, utilizing the same 2 player functionality as Tic-Tac-Toe. This was originally a class assignment, I just refactored it to work with my upgraded Game Console program.

For this game, I had to create DeckOfCards and Card object classes, which can be used for future card games as well. This game should be complete and work. Needs to be tested and debugged.